refactored to use double instead of int

+++ src/shapes/Quadrilateral.java
@@ -1,22 +1,22 @@
 package shapes;
 
 abstract class Quadrilateral extends Shape implements Measurable{
-    protected int length;
-    protected int width;
+    protected double length;
+    protected double width;
 
-    public void setMeasurements(int length, int width){
+    public void setMeasurements(double length, double width){
         this.length = length;
         this.width = width;
     }
 
-    public int getLength() {
+    public double getLength() {
         return this.length;
     }
 
-    public int getWidth() {
+    public double getWidth() {
         return this.width;
     }
 
-    public abstract void setLength(int length);
-    public abstract void setWidth(int width);
+    public abstract void setLength(double length);
+    public abstract void setWidth(double width);
 }
